+++ app/src/main/java/com/example/munchkin/activity/CarddeckActivity.java
@@ -23,7 +23,6 @@ import android.widget.TextView;
 
 import com.example.munchkin.DTO.ActionCardDTO;
 import com.example.munchkin.DrawableUtils.CardUtils;
-import com.example.munchkin.MainGameActivity;
 import com.example.munchkin.MessageFormat.MessageRouter;
 import com.example.munchkin.Player.PlayerHand;
 
@@ -44,9 +43,6 @@ import com.example.munchkin.view.CarddeckView;
 import org.json.JSONException;
 import org.json.JSONObject;
 
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.Collections;
 import java.util.List;
 
 public class CarddeckActivity extends AppCompatActivity {
