+++ app/src/main/java/com/example/munchkin/controller/CardDeckController.java
@@ -1,11 +1,6 @@
 package com.example.munchkin.controller;
 
-import android.widget.ImageView;
-
-import androidx.cardview.widget.CardView;
-
 import com.example.munchkin.DTO.ActionCardDTO;
-import com.example.munchkin.MainGameActivity;
 import com.example.munchkin.MessageFormat.MessageFormatter;
 
 import com.example.munchkin.Player.PlayerHand;
